Refactor plugin structure and add basic environment check.
- The plugin is no longer using a singleton pattern. Instead it is just
a regular class that is being instantiated in the main plugin file.
- Namespace has been added.
- The PHP version check has been added (>= 5.3).
- The helper functions are extracted to a separate file and they are now
using static functions inside the class.
- The `$default_entry_types` array is no longer defined as static.
